AutomotiveRAGFlow/requirements.txt
python-dotenv
pypdf
sentence-transformers
numpy
requests